Make the short int for the BH minimum timeBin just an int, improving alignment

+++ libgadget/blackhole.c
@@ -45,7 +45,7 @@ typedef struct {
     MyFloat BH_MinPotPos[3];
     MyFloat BH_MinPot;
 
-    short int BH_minTimeBin;
+    int BH_minTimeBin;
     MyFloat FeedbackWeightSum;
 
     MyFloat Rho;
